Refactor : "프로젝트 1단계 완료: 폴더 구조 변경"
Domain 패키지 제거후 Entity 패키지 생성
