+++ nakyeonko3/src/main/java/com/group/companytimeclockapp/entity/Employee.java
@@ -1,7 +1,6 @@
-package com.group.companytimeclockapp.domain;
+package com.group.companytimeclockapp.entity;
 
 import com.group.companytimeclockapp.dto.request.EmployeeSaveRequest;
-import com.group.companytimeclockapp.util.Role;
 import jakarta.persistence.*;
 import lombok.Getter;
