+++ nakyeonko3/src/main/java/com/group/companytimeclockapp/entity/Role.java
@@ -1,0 +1,5 @@
+package com.group.companytimeclockapp.entity;
+
+public enum Role {
+    MANAGER, MEMBER
+}
