+++ nakyeonko3/src/main/java/com/group/companytimeclockapp/repository/EmployeeRepository.java
@@ -1,6 +1,6 @@
 package com.group.companytimeclockapp.repository;
 
-import com.group.companytimeclockapp.domain.Employee;
+import com.group.companytimeclockapp.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmployeeRepository extends JpaRepository<Employee,Long> {
