--- nakyeonko3/src/main/java/com/group/companytimeclockapp/util/Role.java
@@ -1,5 +1,0 @@
-package com.group.companytimeclockapp.util;
-
-public enum Role {
-    MANAGER, MEMBER
-}
